Add AbstractServiceFacade with inconsistency log generation logic
