+++ facade/src/main/java/com/redhat/lightblue/migrator/facade/DAOFacadeBase.java
@@ -5,28 +5,14 @@ import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
-import org.json.JSONException;
-import org.reflections.Reflections;
-import org.skyscreamer.jsonassert.JSONCompare;
-import org.skyscreamer.jsonassert.JSONCompareMode;
-import org.skyscreamer.jsonassert.JSONCompareResult;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.ObjectWriter;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.ListeningExecutorService;
 import com.google.common.util.concurrent.MoreExecutors;
@@ -46,23 +32,13 @@ import com.redhat.lightblue.migrator.features.TogglzRandomUsername;
  */
 @SuppressWarnings("all")
 @Deprecated
-public class DAOFacadeBase<D> {
-
-    private static final Logger log = LoggerFactory.getLogger(DAOFacadeBase.class);
-    private static final Logger logInconsisteny = LoggerFactory.getLogger("Inconsistency");
+public class DAOFacadeBase<D> extends AbstractServiceFacade {
 
     protected final D legacyDAO, lightblueDAO;
 
     private EntityIdStore entityIdStore = null;
 
-    private Map<Class<?>,ModelMixIn> modelMixIns;
-
     private int timeoutSeconds = 0;
-
-    private int maxInconsistencyLogLength = 65536; // 64KB
-
-    // used to associate inconsistencies with the service in the logs
-    private final String implementationName;
 
     public EntityIdStore getEntityIdStore() {
         return entityIdStore;
@@ -80,91 +56,11 @@ public class DAOFacadeBase<D> {
     }
 
     public DAOFacadeBase(D legacyDAO, D lightblueDAO) {
-        super();
         this.legacyDAO = legacyDAO;
         this.lightblueDAO = lightblueDAO;
         setEntityIdStore(new EntityIdStoreImpl(this.getClass())); // this.getClass() will point at superclass
-        this.implementationName = this.getClass().getSimpleName();
+        super.implementationName = this.getClass().getSimpleName();
         log.info("Initialized facade for "+implementationName);
-    }
-
-    private Map<Class<?>,ModelMixIn> findModelMixInMappings() {
-        if  (modelMixIns==null) {
-            Reflections reflections = new Reflections("");
-            Set<Class<?>> classes = reflections.getTypesAnnotatedWith(ModelMixIn.class);
-            modelMixIns = new HashMap<>();
-            for (Class<?> clazz : classes) {
-                modelMixIns.put(clazz, clazz.getAnnotation(ModelMixIn.class));
-            }
-        }
-        return modelMixIns;
-    }
-
-    private ObjectWriter getObjectWriter(String methodName) {
-        ObjectMapper mapper = new ObjectMapper();
-        for (Map.Entry<Class<?>, ModelMixIn> entry : findModelMixInMappings().entrySet()) {
-            if (methodName==null || entry.getValue().includeMethods().length==0 || Arrays.asList(entry.getValue().includeMethods()).contains(methodName)) {
-                mapper.addMixInAnnotations(entry.getValue().clazz(), entry.getKey());
-            }
-        }
-        return mapper.writer();
-    }
-
-    // for unit testing
-    public boolean checkConsistency(Object o1, Object o2) {
-        return checkConsistency(o1, o2, null, null);
-    }
-
-    /*
-     * Log message < MAX_INCONSISTENCY_LOG_LENGTH to server.log.
-     * Log message > MAX_INCONSISTENCY_LOG_LENGTH and diff <= MAX_INCONSISTENCY_LOG_LENGTH, log diff to server.log.
-     * Otherwise log method name and parameters to server.log and full message to inconsistency.log.
-     */
-    private void logInconsistency(String callToLogInCaseOfInconsistency, String legacyJson, String lightblueJson, String diff) {
-        String logMessage = String.format("Inconsistency found in %s.%s - diff: %s legacyJson: %s, lightblueJson: %s", implementationName, callToLogInCaseOfInconsistency, diff, legacyJson, lightblueJson);
-        if (logMessage.length()<=maxInconsistencyLogLength) {
-            log.warn(logMessage);
-        } else if (diff!=null&&diff.length()<=maxInconsistencyLogLength) {
-            log.warn(String.format("Inconsistency found in %s.%s - diff: %s", implementationName, callToLogInCaseOfInconsistency, diff));
-        } else {
-            log.warn(String.format("Inconsistency found in %s.%s - payload and diff is greater than %d bytes!", implementationName, callToLogInCaseOfInconsistency, maxInconsistencyLogLength));
-            logInconsisteny.debug(logMessage); // logging at debug level since everything >= info would also land in server.log
-        }
-    }
-
-    public boolean checkConsistency(final Object o1, final Object o2, String methodName, String callToLogInCaseOfInconsistency) {
-        if (o1==null&&o2==null) {
-            return true;
-        }
-
-        String legacyJson=null;
-        String lightblueJson=null;
-        try {
-            long t1 = System.currentTimeMillis();
-            legacyJson = getObjectWriter(methodName).writeValueAsString(o1);
-            lightblueJson = getObjectWriter(methodName).writeValueAsString(o2);
-
-            JSONCompareResult result = JSONCompare.compareJSON(legacyJson, lightblueJson, JSONCompareMode.LENIENT);
-            long t2 = System.currentTimeMillis();
-
-            if (log.isDebugEnabled()) {
-                log.debug("Consistency check took: " + (t2-t1)+" ms");
-                log.debug("Consistency check passed: "+ result.passed());
-            }
-            if (!result.passed()) {
-                logInconsistency(callToLogInCaseOfInconsistency, legacyJson, lightblueJson, result.getMessage().replaceAll("\n", ","));
-            }
-            return result.passed();
-        } catch (JSONException e) {
-            if (o1!=null&&o1.equals(o2)) {
-                return true;
-            } else {
-                logInconsistency(callToLogInCaseOfInconsistency, legacyJson, lightblueJson, null);
-            }
-        } catch (JsonProcessingException e) {
-            log.error("Consistency check failed! Invalid JSON. ", e);
-        }
-        return false;
     }
 
     private ListeningExecutorService createExecutor() {
